Correcion de consultas y reportes

+++ src/main/java/com/example/strorebooks/sales/infraestructure/adapter/in/dto/response/reports/BestSellersByCategoryResponse.java
@@ -16,7 +16,7 @@ public class BestSellersByCategoryResponse {
    List<String> authors;
    String editorial;
    BigDecimal price;
-   BigDecimal sales;
+   Integer sales;
    String category;
    Boolean bestSeller;
 }
